requirements/requirements.txt
bpython==0.23
Django==4.1.1
django-localflavor==3.1
psycopg2-binary
Pillow